+++ japps/PluginLoaders.py
@@ -1,7 +1,6 @@
 from abc import ABCMeta, abstractmethod
 
 from os import PathLike
-from os.path import exists, join
 
 from japps.Configuration import Configuration
 from japps.plugins.IPlugin import IPlugin
@@ -22,7 +21,13 @@ class IPluginLoader(metaclass=ABCMeta):
 class SinglePluginLoader(IPluginLoader):
     def parse_plugin(self) -> IPlugin:
         tmp_plugin = self.config.plugin_class()
+        default_data: dict = self.config.no_info_default.copy()
         parsed_data: dict = OneFileParser.parse(self.path, self.config)
+        for key in default_data.keys():
+            if key in parsed_data.keys():
+                continue
+            parsed_data[key] = default_data[key].format(num=hash(self.path))
+
         for key, value in parsed_data.items():
             setattr(tmp_plugin, key, value)
         if self.config.allow_dependencies and tmp_plugin.DEPENDENCIES:
@@ -34,12 +39,13 @@ class SinglePluginLoader(IPluginLoader):
 class PackagePluginLoader(IPluginLoader):
     def parse_plugin(self) -> IPlugin:
         tmp_plugin = self.config.plugin_class()
-        if exists(join(self.path, self.config.package_plugin_info_filename)):
-            parsed_data: dict = self.config.package_plugin_info_parser.parse(self.path, self.config)
-        else:
-            parsed_data: dict = self.config.no_info_default.copy()
-            for key in parsed_data.keys():
-                parsed_data[key] = parsed_data[key].format(num=hash(self.path))
+
+        default_data: dict = self.config.no_info_default.copy()
+        parsed_data: dict = self.config.package_plugin_info_parser.parse(self.path, self.config)
+        for key in default_data.keys():
+            if key in parsed_data.keys():
+                continue
+            parsed_data[key] = default_data[key].format(num=hash(self.path))
         for key, value in parsed_data.items():
             setattr(tmp_plugin, key, value)